Volume Fraction: Use ROI Manager to limit volume fraction calculations
User is given the option to use the ROI Manager.  If unchecked, ROI Manager is ignored.
Any ROI that is set is applied to every slice in the stack, and if no ROI is set, all voxels in
the stack contribute to BV/TV.

If checked, ROIs relating to each slice are pulled out of the ROI Manager, and voxels are counted in each ROI.  If no ROI is set for the slice, no voxels are counted.  In the case of a single slice image, all ROIs in the ROI Manager are applied to the image, regardless of whether they were originally applied to a particular slice of a multi-slice image.