EllipsoidFactor: display ellipsoid factor stack
EF is calculated as 1 + a/b - b/c where a < b < c and a, b, and c are
the ellipsoid semiaxis lengths (radii). This definition allows
segmentation by thresholding where 0 < ef < 1 are plate-like and 1 < ef
< 2 are rod-like. Values near to 1 are more spheroidal. Subtracting 1
from this definition gives a neater mathematical definition (where -1 is
extreme plate, +1 is extreme rod and 0 is sphere), but this is slighty
unwieldy for thresholding.