Try a rebirth loop
when a vector collides, spawn a single child vector in the face plane
one integer vector back along its length (i.e. in only two of the
dimensions), then add the reborn vector back to the vector set.
Increases sampling density but does not respect occlusion by nearer
background points. Leaks into shadows.

An alternative would be to spawn vectors from the startPoint with the
new vector and iterate until a collision, but how to end the rebirth
loop?